@bradduthie The landscape class is now functioning as expected
landscape class + constructor + create() assigning memory to create a table and fill it with the relevant info + fill() that reset resource level to maximum + getInfo() to cast out the table and check + freeMemory() that free the mem allocated in create()
Small detail: I had to set resource level as an integer for convinience, maybe that's the way to go, we can discuss that this week. Cheers!