All functional and debugged
the seed for the random number generator is now seed in the bash script at every replicate, otherwise the results were the same. The frequency of assessment of the results can now an argument of the model, functional and can be set in the bash script. With current parameters, the predators arrive at the 4th time step, grow and then decline and go extinct while the prey population is still high, important to check if it is not a coding artefact (maybe in the order of the life events?)