Some bugs left, now runs as expected
The max movement range was forced into an int while being a fraction so it was forced to zero and animals were not moving, predators just starved. Fixed now. Some fixes to the simulation script (.sh) in order not to unwillingly delete existing simulation results. @bradduthie Good news ! It's quite clear that with null hypotheses prey populations endure predation similarly, which is what we expected. We can now tweak some parameters to see which creates apparent competition and to which extent.