sim folder for convRate > 1
Hi @bradduthie , 
Hope all is well!
Here is a new simulation folder for the local SA. It will run the SA with values > 1 for the resources/catch ratio between preys, that way we will have an idea of the influence of this parameter on the emergence of apparent competition without the artifact of the maximum number of catches increasing as this ratio decreases.
Could you run this as usual please?
Many thanks,
Adrian